Add utility commands for ping, server, and user
This commit adds three new utility commands: ping, server, and user. Each command is implemented using the SlashCommandBuilder from the discord.js library. The ping command replies with "Pong!", the server command replies with the name of the server where the command was executed, and the user command replies with the tag of the user who executed the command.